[backport from test_harness to tools subtree]
Updated tools to pass -Werror tests.

(cherry picked from commit dd43e06e56fc7e80c667a8eff4cbb13a5040dc55)

+++ algo/numerical.c
@@ -1,9 +1,9 @@
-#include <numerical.h>
+#include "numerical.h"
 #include <math.h>
 
 float nm_fdFirstDer2(const ring_buffer_t r_buff, float dt) {
-  const float A = -1.0;
-  const float B = 1.0;
+  const float A = -1.0f;
+  const float B = 1.0f;
 
   float numerator =
     A*rb_get(r_buff, 1) + //u[n-1]
@@ -13,12 +13,12 @@ float nm_fdFirstDer2(const ring_buffer_t r_buff, float dt) {
 }
 
 float nm_fdFirstDer5(const ring_buffer_t r_buff, float dt) {
-  const float A = -1.0/5.0;
-  const float B = 5.0/4.0;
-  const float C = -10.0/3.0;
-  const float D = 5;
-  const float E = -5;
-  const float F = 137.0/60.0;
+  const float A = -1.0f/5.0f;
+  const float B = 5.0f/4.0f;
+  const float C = -10.0f/3.0f;
+  const float D = 5.0f;
+  const float E = -5.0f;
+  const float F = 137.0f/60.0f;
 
   //ring buffer, 0 is first.
 
@@ -47,12 +47,12 @@ float nm_fdSecondDer3(const ring_buffer_t r_buff, float dt) {
 }
 
 float nm_fdSecondDer5(const ring_buffer_t r_buff, float dt) {
-  const float A = 5.0/6.0;
-  const float B = -61.0/12.0;
-  const float C = 13.0;
-  const float D = -107.0/6.0;
-  const float E = 77.0/6.0;
-  const float F = -15.0/4.0;
+  const float A = 5.0f/6.0f;
+  const float B = -61.0f/12.0f;
+  const float C = 13.0f;
+  const float D = -107.0f/6.0f;
+  const float E = 77.0f/6.0f;
+  const float F = -15.0f/4.0f;
 
   float numerator =
     A*rb_get(r_buff, 5) +
@@ -97,7 +97,7 @@ float nm_sgSecondDer5(const ring_buffer_t r_buff, float dt) {
     E*rb_get(r_buff, 2) +
     G*rb_get(r_buff, 0);
 
-  return numerator/(42.0*dt*dt);
+  return numerator/(42.0f*dt*dt);
 }
 
 float nm_dIntegrate2(const ring_buffer_t prev_vals, float du, float dt) {
@@ -105,10 +105,10 @@ float nm_dIntegrate2(const ring_buffer_t prev_vals, float du, float dt) {
 }
 
 float nm_dIntegrate4(const ring_buffer_t prev_vals, float du, float dt) {
-  const float A = 18.0/11.0;
-  const float B = -9.0/11.0;
-  const float C = 2.0/11.0;
-  const float D = 6.0/11.0;
+  const float A = 18.0f/11.0f;
+  const float B = -9.0f/11.0f;
+  const float C = 2.0f/11.0f;
+  const float D = 6.0f/11.0f;
 
   const float u =
     //this is the previous value since current val [n] not pushed
@@ -121,11 +121,11 @@ float nm_dIntegrate4(const ring_buffer_t prev_vals, float du, float dt) {
 }
 
 size_t nm_genRange(float start, float increment, float end, float **array) {
-  const size_t arr_size = ceil(fabs((end - start)/increment));
+  const size_t arr_size = (size_t) ceil(fabs((end - start)/increment));
   float *range = malloc(arr_size*sizeof(float));
 
   for (size_t i = 0; i < arr_size; ++i) {
-    range[i] = start + i*increment;
+    range[i] = start + ((float)i)*increment;
   }
 
   *array = range;
